girepository: Fix various implicit conversions from size_t to smaller types
Basically various trivial instances of the following MSVC compiler
warning:
```
../gio/gio-tool-set.c(50): warning C4267: '=': conversion from 'size_t' to 'int', possible loss of data
```

+++ girepository/gifieldinfo.c
@@ -190,7 +190,7 @@ gi_field_info_get_field (GIFieldInfo *field_info,
                          void        *mem,
                          GIArgument  *value)
 {
-  int offset;
+  size_t offset;
   GITypeInfo *type_info;
   gboolean result = FALSE;
 
@@ -385,7 +385,7 @@ gi_field_info_set_field (GIFieldInfo      *field_info,
                          void             *mem,
                          const GIArgument *value)
 {
-  int offset;
+  size_t offset;
   GITypeInfo *type_info;
   gboolean result = FALSE;
